+++ src/main/c/psych/extconf.rb
@@ -5,8 +5,18 @@ require 'fileutils'
 
 # :stopdoc:
 
-dir_config 'libyaml'
+if defined?(::TruffleRuby)
+  require 'truffle/libyaml-prefix'
+  dir_config('libyaml', ENV['LIBYAML_PREFIX'])
+else
+  dir_config 'libyaml'
+end
 
+if defined?(::TruffleRuby)
+  # From https://github.com/ruby/psych/blob/v5.0.0/ext/psych/extconf.rb#L42-L43
+  find_header('yaml.h') or abort "yaml.h not found"
+  find_library('yaml', 'yaml_get_version') or abort "libyaml not found"
+else
 if enable_config("bundled-libyaml", false) || !(find_header('yaml.h') && find_library('yaml', 'yaml_get_version'))
   # Embed libyaml since we could not find it.
 
@@ -34,6 +44,7 @@ if enable_config("bundled-libyaml", false) || !(find_header('yaml.h') && find_li
 
   find_header 'yaml.h'
   have_header 'config.h'
+end
 end
 
 create_makefile 'psych' do |mk|